requirements.txt
django
gunicorn
dj-database-url
psycopg2-binary
whitenoise[brotli]
openpyxl
pdf2image
pdfminer.six
pdfplumber
Pillow
PyMuPDF
PyPDF2
pypdfium2
requests
python-telegram-bot
python-dotenv
django-storages
boto3